auto-stop-start/requirements.txt
# Core
boto3>=1.34.0
pyyaml>=6.0.1

# Azure
azure-identity>=1.15.0
azure-mgmt-compute>=30.6.0
azure-mgmt-rdbms>=10.0.0

# Schedule evaluation
croniter>=2.0.5
pytz>=2024.1
